refactor(nlp): helpers modified
