+++ packages/botonic-nlp/tests/helpers/tools-helper.ts
@@ -1,0 +1,4 @@
+import { Dataset } from '../../src/dataset/dataset'
+import { DATA_DIR_PATH } from './constants-helper'
+
+export const dataset = Dataset.load(DATA_DIR_PATH)
